refactor(dejajs-www): 🧩 extract architecture hero primitives to shared module
Move Connector, TapRipple, TypeWriter, SequenceTerminal, SequenceOvalTrack,
SEQ constants, and EASE_ENTER easing out of ArchitectureHero.tsx into a new
animation-primitives.tsx module so the upcoming homepage hero animation can
reuse them. StepLabel stays inlined — only ArchitectureHero uses it.

No behavior change to the /architecture page.
