+++ apps/dejajs-www/components/architecture/ArchitectureHero.tsx
@@ -3,7 +3,14 @@
 import { AnimatePresence, motion } from 'framer-motion';
 import Image from 'next/image';
 import { useEffect, useRef, useState } from 'react';
-import TerminalMockup from './TerminalMockup';
+import {
+  Connector,
+  EASE_ENTER,
+  SEQ,
+  SequenceOvalTrack,
+  SequenceTerminal,
+  TapRipple,
+} from './animation-primitives';
 
 /*
  * ── ARCHITECTURE HERO — TWO-PHASE ANIMATION ──
@@ -18,319 +25,9 @@ import TerminalMockup from './TerminalMockup';
  *   → dot flies conn 3 → "Train moves" label → train orbits! 🚂
  */
 
-const SEQ = {
-  // ── PHASE 1: intro ────────────────────────────────
-  phoneShrinkDur: 0.35,
-  conn1Draw: 0.35,
-  conn1Dur: 0.35,
-  terminalFade: 0.6,
-  terminalDur: 0.2,
-  typingStart: 0.75,    // "deja start" @ 40ms/char ≈ 0.4s
-  readyLine: 1.2,
-  conn2Draw: 1.3,
-  conn2Dur: 0.35,
-  dccFade: 1.55,
-  dccDur: 0.2,
-  conn3Draw: 1.7,
-  conn3Dur: 0.35,
-  trackFade: 1.95,
-  trackDur: 0.2,
-
-  // ── PHASE 2: story (slower — gives viewer time to read each label) ────────
-  tap: 2.25,            // 👆 tap ripple on phone
-  label1: 2.35,         // "You tap" / "Throttle App"
-  dot1: 2.95,           // pause 0.6s after label1 so the viewer can read it
-  dotDur: 0.8,
-  label2: 3.85,         // "Server routes" / "DEJA Server" — as dot arrives
-  throttleCmd: 4.35,    // ← throttle line
-  serialCmd: 4.7,       // → <t 3 50 1> line
-  dot2: 5.25,
-  label3: 6.15,         // "Translates & transmits" / "DCC-EX CommandStation"
-  dot3: 6.65,
-  label4: 7.55,         // "Train moves" / "DCC Track"
-  trainStart: 7.65,     // 🚂 train begins orbit
-} as const;
-
-const EASE_ENTER = [0.22, 1, 0.36, 1] as [number, number, number, number];
-
 interface ArchitectureHeroProps {
   /** Start the sequence immediately on mount (used by the replay button). */
   autoStart?: boolean;
-}
-
-/** Smooth cable-drape S-curve connector, optional traveling command dot. */
-function Connector({
-  direction,
-  drawDelay,
-  drawDuration = 0.35,
-  color1,
-  color2,
-  id,
-  showDot = false,
-  dotDelay = 0,
-  dotDuration = 0.6,
-}: {
-  direction: 'left-to-right' | 'right-to-left';
-  drawDelay: number;
-  drawDuration?: number;
-  color1: string;
-  color2: string;
-  id: string;
-  showDot?: boolean;
-  dotDelay?: number;
-  dotDuration?: number;
-}) {
-  // Original vertical-tangent S — renders as a letterboxed square in the row
-  // center thanks to the default preserveAspectRatio="xMidYMid meet". Graphics
-  // are positioned near the row center to sit close to the curve endpoints.
-  const d =
-    direction === 'left-to-right'
-      ? 'M 10,2 C 10,40 90,60 90,98'
-      : 'M 90,2 C 90,40 10,60 10,98';
-
-  return (
-    <div className="w-full h-10 sm:h-12">
-      <svg
-        viewBox="0 0 100 100"
-        className="w-full h-full"
-        fill="none"
-      >
-        <path
-          d={d}
-          stroke={color1}
-          strokeOpacity={0.06}
-          strokeWidth={2}
-          strokeLinecap="round"
-          vectorEffect="non-scaling-stroke"
-        />
-        <motion.path
-          d={d}
-          stroke={`url(#${id})`}
-          strokeWidth={2}
-          strokeLinecap="round"
-          vectorEffect="non-scaling-stroke"
-          initial={{ pathLength: 0 }}
-          animate={{ pathLength: 1 }}
-          transition={{ duration: drawDuration, delay: drawDelay, ease: EASE_ENTER }}
-        />
-        {showDot && (
-          <>
-            <motion.circle
-              r={4}
-              fill={color1}
-              initial={{ opacity: 0 }}
-              animate={{ opacity: [0, 1, 1, 0] }}
-              transition={{ delay: dotDelay, duration: dotDuration, times: [0, 0.05, 0.9, 1] }}
-            >
-              <animateMotion dur={`${dotDuration}s`} fill="freeze" begin={`${dotDelay}s`}>
-                <mpath href={`#${id}-path`} />
-              </animateMotion>
-            </motion.circle>
-            <motion.circle
-              r={8}
-              fill={color1}
-              initial={{ opacity: 0 }}
-              animate={{ opacity: [0, 0.25, 0.25, 0] }}
-              transition={{ delay: dotDelay, duration: dotDuration, times: [0, 0.05, 0.9, 1] }}
-            >
-              <animateMotion dur={`${dotDuration}s`} fill="freeze" begin={`${dotDelay}s`}>
-                <mpath href={`#${id}-path`} />
-              </animateMotion>
-            </motion.circle>
-          </>
-        )}
-        <path id={`${id}-path`} d={d} fill="none" stroke="none" />
-        <defs>
-          <linearGradient id={id} x1="0%" y1="0%" x2="0%" y2="100%">
-            <stop offset="0%" stopColor={color1} stopOpacity={0.8} />
-            <stop offset="100%" stopColor={color2} stopOpacity={0.6} />
-          </linearGradient>
-        </defs>
-      </svg>
-    </div>
-  );
-}
-
-/** 👆 Tap ripple on the phone. */
-function TapRipple({ delay }: { delay: number }) {
-  return (
-    <motion.div
-      className="absolute inset-0 flex items-center justify-center pointer-events-none z-20"
-      initial={{ opacity: 0 }}
-      animate={{ opacity: [0, 1, 0] }}
-      transition={{ delay, duration: 0.7 }}
-    >
-      <motion.div
-        className="w-8 h-8 rounded-full bg-white/30 backdrop-blur-sm"
-        initial={{ scale: 0.5 }}
-        animate={{ scale: [0.5, 1, 0.8] }}
-        transition={{ delay, duration: 0.3 }}
-      />
-      <motion.div
-        className="absolute w-16 h-16 rounded-full border-2 border-deja-cyan/50"
-        initial={{ scale: 0.5, opacity: 0 }}
-        animate={{ scale: [0.5, 2.0], opacity: [0, 0.7, 0] }}
-        transition={{ delay: delay + 0.05, duration: 0.5 }}
-      />
-    </motion.div>
-  );
-}
-
-/** Types text one character at a time after an absolute delay. */
-function TypeWriter({
-  text,
-  startAt,
-  speed = 40,
-}: {
-  text: string;
-  startAt: number;
-  speed?: number;
-}) {
-  const [count, setCount] = useState(0);
-
-  useEffect(() => {
-    let interval: ReturnType<typeof setInterval> | null = null;
-    const startTimer = setTimeout(() => {
-      interval = setInterval(() => {
-        setCount((c) => {
-          if (c >= text.length) {
-            if (interval) clearInterval(interval);
-            return c;
-          }
-          return c + 1;
-        });
-      }, speed);
-    }, startAt * 1000);
-
-    return () => {
-      clearTimeout(startTimer);
-      if (interval) clearInterval(interval);
-    };
-  }, [text, startAt, speed]);
-
-  return <span>{text.slice(0, count)}</span>;
-}
-
-/** Terminal with typewriter prompt + rolling output lines. */
-function SequenceTerminal() {
-  // Delays are relative to terminal mount, so subtract terminalFade offset.
-  const typingStart = SEQ.typingStart - SEQ.terminalFade;
-  const readyStart = SEQ.readyLine - SEQ.terminalFade;
-  const throttleStart = SEQ.throttleCmd - SEQ.terminalFade;
-  const serialStart = SEQ.serialCmd - SEQ.terminalFade;
-
-  return (
-    <TerminalMockup title="deja — tamarack" className="text-[10px] w-[210px] sm:w-[240px]">
-      <p className="text-deja-cyan">
-        $ <TypeWriter text="deja start" startAt={typingStart} />
-      </p>
-      <motion.p
-        className="text-green-400"
-        initial={{ opacity: 0, y: 4 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ delay: readyStart, duration: 0.18, ease: 'easeOut' }}
-      >
-        ✓ Ready
-      </motion.p>
-      {/* 🏗️ reserved space — actual content rolls in during phase 2 */}
-      <p aria-hidden className="invisible">{'\u00A0'}</p>
-      <motion.p
-        className="text-yellow-400"
-        initial={{ opacity: 0, y: 4 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ delay: throttleStart, duration: 0.25, ease: 'easeOut' }}
-      >
-        ← throttle {'{addr:3, speed:50}'}
-      </motion.p>
-      <motion.p
-        className="text-deja-magenta"
-        initial={{ opacity: 0, y: 4 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ delay: serialStart, duration: 0.25, ease: 'easeOut' }}
-      >
-        → {'<t 3 50 1>'}
-      </motion.p>
-    </TerminalMockup>
-  );
-}
-
-/** Stadium track — train parked initially, starts orbiting at `trainStartDelay`. */
-function SequenceOvalTrack({ trainStartDelay }: { trainStartDelay: number }) {
-  const outerTrack = 'M 60,10 L 140,10 A 30,30 0 0 1 140,70 L 60,70 A 30,30 0 0 1 60,10 Z';
-  const innerTrack = 'M 62,14 L 138,14 A 26,26 0 0 1 138,66 L 62,66 A 26,26 0 0 1 62,14 Z';
-  const orbitPath = 'M 61,12 L 139,12 A 28,28 0 0 1 139,68 L 61,68 A 28,28 0 0 1 61,12 Z';
-
-  const [moving, setMoving] = useState(false);
-
-  useEffect(() => {
-    const timer = setTimeout(() => setMoving(true), trainStartDelay * 1000);
-    return () => clearTimeout(timer);
-  }, [trainStartDelay]);
-
-  const ties: { x1: number; y1: number; x2: number; y2: number }[] = [];
-  for (let i = 0; i < 8; i++) { const x = 65 + i * 10; ties.push({ x1: x, y1: 7, x2: x, y2: 17 }); }
-  for (let i = 0; i < 8; i++) { const x = 65 + i * 10; ties.push({ x1: x, y1: 63, x2: x, y2: 73 }); }
-  for (let i = 0; i < 5; i++) {
-    const a = -Math.PI / 2 + (i + 1) * (Math.PI / 6);
-    ties.push({
-      x1: Math.round(140 + 25 * Math.cos(a)),
-      y1: Math.round(40 + 25 * Math.sin(a)),
-      x2: Math.round(140 + 35 * Math.cos(a)),
-      y2: Math.round(40 + 35 * Math.sin(a)),
-    });
-  }
-  for (let i = 0; i < 5; i++) {
-    const a = Math.PI / 2 + (i + 1) * (Math.PI / 6);
-    ties.push({
-      x1: Math.round(60 + 25 * Math.cos(a)),
-      y1: Math.round(40 + 25 * Math.sin(a)),
-      x2: Math.round(60 + 35 * Math.cos(a)),
-      y2: Math.round(40 + 35 * Math.sin(a)),
-    });
-  }
-
-  return (
-    <div className="w-[140px] sm:w-[170px] h-[56px] sm:h-[68px] relative">
-      <svg viewBox="0 0 200 80" className="w-full h-full" fill="none">
-        {ties.map((t, i) => (
-          <line
-            key={i}
-            x1={t.x1}
-            y1={t.y1}
-            x2={t.x2}
-            y2={t.y2}
-            stroke="#F97316"
-            strokeOpacity={0.2}
-            strokeWidth={1.5}
-          />
-        ))}
-        <path d={outerTrack} stroke="#F97316" strokeOpacity={0.35} strokeWidth={1.2} />
-        <path d={innerTrack} stroke="#F97316" strokeOpacity={0.25} strokeWidth={1} />
-
-        {moving ? (
-          // 🚂 train chain orbiting the path
-          [0, 1, 2, 3, 4].map((i) => (
-            <circle key={i} r={i === 0 ? 4.5 : 3} fill="#F97316" fillOpacity={i === 0 ? 1 : 0.75 - i * 0.1}>
-              <animateMotion dur="3s" repeatCount="indefinite" begin={`${i * 0.1}s`}>
-                <mpath href="#stadium-orbit" />
-              </animateMotion>
-            </circle>
-          ))
-        ) : (
-          // 🛑 train parked at the top-left
-          <>
-            <circle cx={61} cy={12} r={4.5} fill="#F97316" />
-            <circle cx={71} cy={12} r={3} fill="#F97316" fillOpacity={0.75} />
-            <circle cx={81} cy={12} r={3} fill="#F97316" fillOpacity={0.65} />
-            <circle cx={91} cy={12} r={3} fill="#F97316" fillOpacity={0.55} />
-            <circle cx={101} cy={12} r={3} fill="#F97316" fillOpacity={0.45} />
-          </>
-        )}
-
-        <path id="stadium-orbit" d={orbitPath} fill="none" stroke="none" />
-      </svg>
-    </div>
-  );
 }
 
 /**
